# frontend/openapi.core.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/portfolio/summary": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get portfolio summary",
        "description": "Returns aggregated portfolio statistics including total value, returns, and changes",
        "operationId": "getPortfolioSummary",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved portfolio summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioSummaryResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/portfolio/performers": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get top and bottom performers",
        "description": "Returns top and bottom performing assets with their percentage change",
        "operationId": "getTopBottomPerformers",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5,
              "minimum": 1
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved performers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopBottomPerformersResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid limit parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopBottomPerformersResponse"
                }
              }
            }
          }
        }
      }
    },
    "/portfolio/performance-data": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get combined performance data",
        "description": "Returns the portfolio chart series and monthly invested amounts together",
        "operationId": "getPortfolioPerformanceData",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved performance data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioPerformanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioPerformanceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/portfolio/monthly-invested": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get monthly invested amounts",
        "description": "Returns invested euros per month for the last N months",
        "operationId": "getMonthlyInvested",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved monthly invested amounts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyInvestedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/portfolio/monthly-dividends": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get monthly dividends",
        "description": "Returns dividends per month for the last N months (null = entire history)",
        "operationId": "getMonthlyDividends",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved monthly dividends"
          }
        }
      }
    },
    "/portfolio/holdings": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get current portfolio holdings",
        "description": "Returns a list of all current portfolio positions with detailed information.",
        "operationId": "getHoldings",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved holdings list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetHoldingResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/portfolio/holdings/summary": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get active positions summary",
        "description": "Returns active positions with invested amount, invested %, P/L (EUR and %), market value and market value %",
        "operationId": "getActivePositions",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved positions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivePositionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/portfolio/dividend-cost-yield": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get dividend cost yield",
        "description": "Returns dividend cost yield for each asset and total portfolio. Yield is calculated as (Annualized Dividend / Cost Basis) * 100, annualized from first purchase date.",
        "operationId": "getDividendCostYield",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved dividend cost yield",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DividendCostYieldResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/portfolio/consolidated-holdings": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get consolidated portfolio holdings by company or ETF",
        "description": "Returns a list of all current portfolio positions grouped by company or ETF with profit/loss and dividend yield.",
        "operationId": "getConsolidatedHoldings",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved consolidated holdings list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyAssetHoldingResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/portfolio/concentration": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get portfolio concentration and HHI",
        "description": "Returns top1/top3/top5/top10 concentration percentages and Herfindahl–Hirschman Index (HHI)",
        "operationId": "getPortfolioConcentration",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved concentration metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioConcentrationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/portfolio/chart/performance": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get portfolio value time series",
        "description": "Returns daily market value and invested amount for the specified time period, optionally including benchmark indices",
        "operationId": "getPortfolioPerformanceChart",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved value series",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioChartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid days parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioChartResponse"
                }
              }
            }
          }
        }
      }
    },
    "/portfolio/allocation": {
      "get": {
        "tags": [
          "Portfolio Controller"
        ],
        "summary": "Get combined allocation",
        "description": "Returns allocations by currency, industry/sector and country together",
        "operationId": "getCombinedAllocation",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TRADING212",
                "ETORO",
                "IBKR",
                "REVOLUT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved allocation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioAllocationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PortfolioSummaryResponse": {
        "type": "object",
        "description": "Portfolio summary with total value and returns",
        "properties": {
          "snapshotDate": {
            "type": "string",
            "format": "date"
          },
          "totalValue": {
            "type": "number"
          },
          "totalInvested": {
            "type": "number"
          },
          "totalReturn": {
            "type": "number"
          },
          "totalReturnPercentage": {
            "type": "number"
          },
          "totalUnrealizedPL": {
            "type": "number"
          },
          "totalRealizedPL": {
            "type": "number"
          },
          "totalHoldings": {
            "type": "integer",
            "format": "int32"
          },
          "totalDividends": {
            "type": "number"
          },
          "dividendCostYield": {
            "type": "number"
          },
          "periodChange": {
            "$ref": "#/components/schemas/PeriodChangeResponse"
          }
        }
      },
...
      "PerformerResponse": {
        "type": "object",
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetSimpleResponse"
          },
          "percentageChange": {
            "type": "number"
          }
        }
      },
      "TopBottomPerformersResponse": {
        "type": "object",
        "properties": {
          "top": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PerformerResponse"
            }
          },
          "bottom": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PerformerResponse"
            }
          }
        }
      },
...
      "MonthlyInvestedResponse": {
        "type": "object",
        "properties": {
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyInvestedEntry"
            }
          }
        }
      },
...
      "PortfolioPerformanceResponse": {
        "type": "object",
        "properties": {
          "chart": {
            "$ref": "#/components/schemas/PortfolioChartResponse"
          },
          "monthlyInvested": {
            "$ref": "#/components/schemas/MonthlyInvestedResponse"
          }
        }
      },
...
    }
  }
}